Corrected calculation of various terms.
Removed use of curl and divergence functions in MATLAB, only now
using gradient function and then combinations of resulting terms.
MATLAB treats x coordinates as y coordinates and this makes things
confusing. We know how to fix this for gradient, but it's unclear
for the other functions, so the codes now just use gradient calculations
only.