Recalculated everything after realizing MATLAB issues.
MATLAB swaps the x and y axes of matrices when using things like
gradient, divergence, and curl functions... so lots was wrong.
The updated codes now only use the gradient() function and
determine all other terms based on output from that, with the
appropriate swapping of x and y terms in the output.